+++ cpp/src/core/cpu_operations.cc
@@ -1,4 +1,4 @@
-#include "CpuOperations.h"
+#include "cpu_operations.h"
 
 namespace nice {
 template <typename T>
