+++ cpp/src/core/gpu_operations.h
@@ -20,28 +20,34 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#include "Util.h"
-#include <fstream>
-#include <iostream>
+#ifndef CPP_SRC_CORE_GPU_OPERATIONS_H_
+#define CPP_SRC_CORE_GPU_OPERATIONS_H_
+
+#include "matrix.h"
 
 namespace nice {
 
-template<typename T>
-Matrix<T> Util<T>::FromFile(std::string input_file_path) {
+// Abstract class of common matrix operation interface
+template <typename T>
+class GpuOperations {
+ public:
+  static Matrix<T> Multiply(const Matrix<T> &a, const T &scalar);
+  static Matrix<T> Multiply(const Matrix<T> &a, const Matrix<T> &b);
+  static Matrix<T> Add(const Matrix<T> &a, const T &scalar);
+  static Matrix<T> Add(const Matrix<T> &a, const Matrix<T> &b);
+  static Matrix<T> Subtract(const Matrix<T> &a, const T &scalar);
+  static Matrix<T> Subtract(const Matrix<T> &a, const Matrix<T> &b);
+  static Matrix<T> Inverse(const Matrix<T> &a);
+  static Matrix<T> Norm(const int &p = 2, const int &axis = 0);
+  static T Determinant(const Matrix<T> &a);
+  static T Rank(const Matrix<T> &a);
+  static T FrobeniusNorm(const Matrix<T> &a);
+  static T Trace(const Matrix<T> &a);
+  static T DotProduct(const Vector<T> &a, const Vector<T> &b);
+  static Matrix<T> OuterProduct(const Vector<T> &a, const Vector<T> &b);
+};
 
-
-}
-
-
-//}
-//  std::ifstream input_file(input_file_path);
-//  if (input_file) {
-////    for (int i = 0; i < num_rows_; i++)
-////      for (int j = 0; j < num_cols_; j++)
-////        input_file >> (*matrix_)(i, j);
-//    input_file.close();
-//    return true;
-//  } else
-//    return false;
-//}
 }  // namespace nice
+
+#endif  // CPP_SRC_CORE_GPU_OPERATIONS_H_
+
