+++ cpp/src/core/matrix.h
