+++ cpp/src/core/vector.h
